refactor: extract TopContributor into shared component
- Create shared TopContributor component in src/components/nostr/TopContributor.tsx
- Support two variants: "default" (larger, shows "sats" suffix) and "compact" (smaller, no suffix)
- Update user menu to use shared component with default variant
- Update welcome screen to use shared component with compact variant
- Remove duplicate code and consolidate formatting logic
- Maintain consistent visual design across both usages
