+++ src/components/GrimoireWelcome.tsx
@@ -1,12 +1,11 @@
-import { Terminal, Trophy } from "lucide-react";
+import { Terminal } from "lucide-react";
 import { Button } from "./ui/button";
 import { Kbd, KbdGroup } from "./ui/kbd";
 import { Progress } from "./ui/progress";
 import supportersService, { MONTHLY_GOAL_SATS } from "@/services/supporters";
 import { useLiveQuery } from "dexie-react-hooks";
 import db from "@/services/db";
-import { useProfile } from "@/hooks/useProfile";
-import { getDisplayName } from "@/lib/nostr-utils";
+import { TopContributor } from "./nostr/TopContributor";
 
 interface GrimoireWelcomeProps {
   onLaunchCommand: () => void;
@@ -29,36 +28,6 @@ const EXAMPLE_COMMANDS = [
   },
   { command: "req -k 1 -l 20", description: "Query recent notes" },
 ];
-
-function TopContributor({
-  pubkey,
-  amount,
-}: {
-  pubkey: string;
-  amount: number;
-}) {
-  const profile = useProfile(pubkey);
-  const displayName = getDisplayName(pubkey, profile);
-
-  function formatNumber(sats: number): string {
-    if (sats >= 1_000_000) {
-      return `${(sats / 1_000_000).toFixed(1)}M`;
-    } else if (sats >= 1_000) {
-      return `${Math.floor(sats / 1_000)}k`;
-    }
-    return sats.toString();
-  }
-
-  return (
-    <div className="flex items-center gap-1.5 mt-1.5 pt-1.5 border-t border-border/30">
-      <Trophy className="size-3 text-yellow-500" />
-      <span className="text-[10px] text-muted-foreground flex-1 truncate">
-        {displayName}
-      </span>
-      <span className="text-[10px] font-medium">{formatNumber(amount)}</span>
-    </div>
-  );
-}
 
 export function GrimoireWelcome({
   onLaunchCommand,
@@ -187,6 +156,7 @@ export function GrimoireWelcome({
                     <TopContributor
                       pubkey={topContributor.pubkey}
                       amount={topContributor.totalSats}
+                      variant="compact"
                     />
                   )}
                 </div>
